+++ src/basic/components/AddToCartButton.ts
@@ -1,7 +1,7 @@
-import { createElement } from '../utils/createElement';
+import { createElement as h } from '../utils/createElement';
 
 export const AddToCartButton = () => {
-  return createElement('button', {
+  return h('button', {
     id: 'add-to-cart',
     className: 'bg-blue-500 text-white px-4 py-2 rounded',
     textContent: '추가',
