+++ src/basic/services/setCartTotalText.ts
@@ -1,0 +1,3 @@
+export const updateCartTotalText = ($cartTotal: HTMLElement, totalAmount: number) => {
+  $cartTotal.textContent = `총액: ${totalAmount}원`;
+};
